Change exitCode from Integer to Long
This is a continuation of https://github.com/spotify/docker-client/pull/1053 as parts of docker-client
were still using exit code of Integer. This caused some uses to have inconsistent types.

+++ src/main/java/com/spotify/docker/client/messages/ContainerExit.java
@@ -34,10 +34,10 @@ import com.google.auto.value.AutoValue;
 public abstract class ContainerExit {
 
   @JsonProperty("StatusCode")
-  public abstract Integer statusCode();
+  public abstract Long statusCode();
 
   @JsonCreator
-  public static ContainerExit create(@JsonProperty("StatusCode") final Integer statusCode) {
+  public static ContainerExit create(@JsonProperty("StatusCode") final Long statusCode) {
     return new AutoValue_ContainerExit(statusCode);
   }
 }
